+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+* Collect the principal variation during search (~18 Elo)
 * Add aspiration windows (~5 Elo)
 * Return `best_eval` in quiescence (~5 Elo)
 * Fail soft on TT cuts (~2 Elo)